(#4855) Fix group resource in OS X
The group provider on OS X uses “dseditgroup” to manage group
membership. Due to Apple bug 8481241 (“dseditgroup can’t remove unknown
users from groups”), however, if the puppet group provider needs to
remove a non-existant user from a group it manages, it will fail.

To remedy this, in the meantime, the provider will call dscl to delete
the non-existant member from the group. If that fails then the error
is rescued and feedback is provided.