(#11803) Add uninstall action for the module face
Without this patch, the module face does not have an action for
uninstalling Puppet modules.

This patch adds an uninstall action to the module face, with initial
support for uninstalling Puppet modules from the default module paths or
specific directories.

This patch also includes the related spec tests.

Example Usage:
Running the following commands will uninstall a module from all
module paths or a specific directory

    $ puppet module uninstall apache
    Removed /etc/puppet/modules/apache

    $ puppet module uninstall --target-directory /usr/share/puppet/modules apache
    Removed /usr/share/puppet/modules/apache

Running the Tests:
The uninstall spec tests can be run with the following commands:

    $ rspec spec/unit/face/module/uninstall_spec.rb
    $ rspec spec/unit/module_tool/uninstaller_spec.rb