(#12349) Add module dependency errors to module list output
There was already a validate_dependencies method on the Puppet::Module
class, but it didn't do anything since it never actually read the
dependency info from the module metadata.

This commit adds a method to replace that, but it doesn't raise if
dependencies are unsatisfied.  This allows us to output the dependency
problems when `puppet module list` is called.

This also adds basic acceptance tests for the list command.