Ready repository for release process
This commit modifies the behavior of the Rake task to use the tagged
version as the version of the repository/gem.  This will allow for
autobuilding in the future.  This is extremely similar to the heira
Rakefile and tasks.

Also got rid of a package_task warning in the Rakefile.

I also tagged the original commit at v0.0.0 to ensure this process would
work.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>