(#11929) Always serve files in binary mode
Previously, Windows agents were reading files in text mode when serving
them locally, such as when serving files from a local module, corrupting
binary files in the process.

This commit reads files in binary mode, which is a noop on Unix.

Serving files from remote puppet masters was fixed in #9983.