(#11888) Revert 483a1d9 for 2.7.x only
This reverts commit 483a1d93a5e5c78f8f41f905d7aae896081c98f1, reversing
changes made to c6667c50d3b49195685311575b46de1978c4dfd9.

This revert is necessary because the change set includes API breaking
changes in a patch version release of the product.

The API breaking changes are that Puppet 2.7.x > 2.7.9 would require a
different version of Facter than 2.7.x <= 2.7.9 if this change set were
to be released.

This commit should be reverted after 2.7.x merges up into master.  That
is to say, the revert itself should be reverted.