(#11996) Fix test failures due to hash processing order changes.
As part of fixing CVE-2011-4815, Ruby has made the order that hash tables are
keyed unpredictable - where they were formerly fixed in stone.  This, in turn,
means that all the tests that depended on that order started to fail randomly
on the most recent patch release of Ruby 1.8.7, and will fail in recent Ruby
1.9.* releases.

This makes those tests insensitive to ordering of the content, at the cost of
being a little less strict in places.  This seems less harmful, overall, than
alternative approaches like sorting the hash in the code, or enumerating all
possible combinations of order and testing on all of them.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>