(#11996) Fix file content test after hash changes.
CVE-2011-4815 changed the order that hashes returned their content; that, in
turn, changed the order that directory entries were returned to something much
more likely to be random.

This broke the test that assumed the order of '.' and 'file.rb' would be
consistent; this fixes that by finding the right entry in the result instead
of assuming it is always in the same slot.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>