(#11600) Remove module face
This commit removes the face related code and tests since it was
determined they didn't meet design specifications well enough for the
upcoming release. This leaves the module_tool related code in, but it's
not accessible through any supported interface. This will make it easy
for us to get this removed code back in after the release and continue
development for the next release.