(#11057) Add spec, should notify why a run is skipped
Without this patch, there isn't any spec test coverage for this
particular bug and the behavior change introduced in this change set.

This patch fixes the problem by adding a spec test covering our
expectation that Puppet informs the user why a run is skipped.

The test makes the fundamental assumption this information is contained
within Puppet::Application.run_status which we're mocking.  The test
validates the information contained in
Puppet::Application.run_status is in fact shown to the user.