(#11727) Allocate a PTY when running the acceptance test.
When submitted this acceptance test didn't actually pass - which, ultimately,
turns out to be that the puppet command was run without a TTY attached by
default in the acceptance test harness.

Once https://github.com/puppetlabs/puppet-acceptance/pull/115 is merged, which
adds the feature to the core of the suite to support PTY allocation through
SSH, this turns that feature on, and the test can finally pass.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>