(maint) Fix test 3360 to use per host cfg data
Test harness formerly used a global config hash to store config data;
this breaks with the intoduction on non-posix hosts (windows).  Now
using a per host config hash.  Updating this test to use the new
style.  This will need to be merged into 2.7rc as well; already
applied to Master.