(#11858) Don't load plugins during sync.
Previous, any ruby file downloaded during pluginsync would be loaded. This
causes inconsistent behavior, because this only happened on runs when the
files needed to be updated (e.g. they weren't present on the agent before).
Removing this means only files that are autoloaded by some other part of the
code get loaded, which was the only case that worked consistently.