(#6710) Fix Ruby 1.9.2 failures due to Array#to_s changes.
The behaviour of Array#to_s is different between 1.8.7 and 1.9.2, and this
code actually depended on the older behaviour.

Since this is part of a stable series, the smallest set of changes possible to
make things work have been applied - and a ticket filed to do the deep
analysis of data flow required to fix this properly in a release where we have
more stomach for risk.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>