(#10296) Switch schedule tests from should to must
Per Patrick's explanation in #10296, all of the
"should" forms of the rspec tests have been converted
to "must" forms to that the tests will function properly.
As part of this cleanup, one test was found that was "should"
and converted to "must_not" as the original logic seemed
inconsistent with the test.