(#12080) Add custom console logging for Faces
Without this patch, Face commands do not support custom formatting of
console output.

This patch fixes this issue by enhancing the current console logging
destination with support for custom formatting of messages that
originate from `Puppet::Interface`.

This patch enables the ability to log warnings, errors, and notices to
stderr with better formatting and colors:

    Puppet::Face.err "This error will go to stderr with nice formatting"

This patch enables the ability to color specific sections of a given string:

    "This is a red #{Puppet::Face.colorize(:red, 'string'}"

Related spec tests are included.