(#12088) Add windows:build task to build MSI package
Without this patch, we're not actually building anything yet.  This
patch fixes that problem by exercising the full process of downloading
Ruby, unpacking it into a staging area, then building an MSI suitable to
install Ruby into C:\puppetlabs\sys\ruby\bin\ruby.exe

Running `rake windows:build` will produce puppet.msi.  Puppet is not actually
included in the installation package, but Ruby is and subsequent patches
will add the functionality to include puppet and facter.

This patch is a migration of (#11890) into the puppet project itself.  The goal
is to be able to use Puppet's 2.7.x branch to build packages for arbitrary
integrations of Puppet, Facter, and other repositories going into the Windows
MSI package.