(#11860) Add parameter to control locale override
Puppet::Util.execute_posix contained some code that would override some locale-related environment variables in order to force the command being executed to produce well-known, consistent, predictably formatted output so that it could be parsed.  However, there are some cases (such as a user-defined "exec" block) where we need to respect the system/user locale settings.  Add a boolean parameter to the Puppet::Util.execute method that allows toggling this locale override behavior.

The "exec" provider now passes "false" for this value, meaning that we no longer override the locale for user-defined "exec" resources.  Users may choose to override the locale themselves by specifying the appropriate locale-related vars in the "environment" section of their resource.

All of this is ignored on windows at the moment, but the parameter is passed in to "execute", which is responsible for calling "execute_windows" on windows systems.  We could decide to do something with this parameter on windows at some point in the future.