(#12066) Add puppet and facter to puppet.msi
Without this patch only the Ruby runtime environment is bundled into the
puppet.msi package.  This patch builds upon this work adding puppet and
facter from Git source into the puppet.msi package.

This work is not yet complete in that installing puppet.msi does not
yield a system where `puppet agent --test` works out of the box, but it
gets us significantly closer to the goal.

In particular, the versions of Puppet and Facter are parameterized by
this patch.  If the windows:checkout task is not explicitly called prior
to calling the windows:build task then the build system will default to
checking out Puppet 2.7.9 and Facter 1.6.4.

To build a specific version:

    rake windows:checkout[refs/tags/2.7.10,refs/tags/1.6.4]
    rake clean
    rake windows:build

To build a specific integration branch:

    rake windows:checkout[refs/remotes/origin/2.7.x,refs/remotes/origin/1.6.x]
    rake clean
    rake windows:build

To build from some other repository:

    rake clobber
    rake windows:clone[git://github.com/jeffmccune/puppet.git]
    rake windows:checkout[refs/remotes/origin/feature/2.7.x/12066_myfeature]
    rake clean
    rake windows:build