(#12169) Correct redhat spec changelog chronology issue
Without this patch, two entries in the %changelog section of the
puppet.spec file in conf/redhat are out of chronological order,
due to merge-ups from 2.6.x. This causes the rpm to not build
cleanly. This patch removes the two entries that are out of order
(placing them in correct date chronology would have resulted
in the version numbers being out of order, hence the removal).

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>