Revert "Finer-grained protocol for property `insync?`."
This reverts commit 48726b662d2fb5cc3532fca12eb6aa5a3c1622cd.

Revert the commit to use the old insync? method. Revert the commit
due to the following different reasons:

* the code does only work on ruby 1.9 as described in #12197
* the insync? method does not work for array comparisons if
  the should array has more items than the is array
  because smallarray.zip(bigarray) truncates bigarray
* it is currently not possible to overwrite just the property_match?
  method because the insync method will then raise an exception
  Could not evaluate: old and new mismatch!

While I guess there were good reasons behind the refactor of the
insync? method it is currently quicker to just revert the commit than
fixing the errors (and writing tests for it)