Property Spec cleanup: extract new subclass to let method.
We need a shiny, new subclass of Puppet::Property every test, because we
modify class level things - and so we would pollute the tests if we touched
the same class object.

This extracts the logic around that to a let method, making the lifecycle
easier to follow by eliminating a member variable.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>