Property Spec cleanup: remove some pointless extra stubs.
One test context had a duplicate version of the outer instance management
code, probably because it was in a different class so couldn't see the member.

Replacing the outer version with let methods means that we also see the
helpers down in our nested contexts, so we can ditch the extra code.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>