Add unit tests for the `insysc?` method of a property.
Previously this core, vital method was covered by a bunch of other random
stuff, including acceptance tests, but it didn't actually have tests that
belonged to just it.

This was revealed to be a problem in the last refactor of the code, which blew
up in nasty ways when an edge case wasn't found.

These tests do, in fact, show up that problem and everything.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>