Restore compatible `insync?` behaviour for matching arrays.
Prior to the start of these cleanups, when matching the entire array rather
than part of it, we accepted the situation where the `is` array - what the
provider returned - matched one of two arrays:

One, if it matched as-is the `should` array, which is what the resource is
configured to be delivering.

Two, if it matched the `should` array after every value in it was transformed
via `to_s`, into a string.

Preserving this does mean that the `property_matches?` method is no longer
used for comparison when `array_matching :all` is in force - because it
internally handles the value comparisons, we can't expose the logic that
refuses to match anything but the entire array to it.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>