Revert "Merge pull request #396 from jeffmccune/feature/2.7.x/12088_windows_build_tasks"
This reverts commit dd6bef5446a3ce892b503ac19830f810791560ad, reversing
changes made to 586074aca8dd49843ebfb9f160b408b7907cf886.

We've decided to back this out because the windows build system packages
Ruby and Facter in addition to Puppet.  Therefore, the windows build
system should be in a separate repository.

Please see http://projects.puppetlabs.com/issues/11205 for the location
of the new repository.