Alternate, and portable, stubbing of execution for upstart spec.
Previously this spec used to execute an external command in order to set the
Ruby `$?` global to a `Process::Status` object with a successful exit.  This
was fine, but not portable to Windows, and not exactly efficient.

Instead, we now stub the exitcode on the $? object, whatever it is, to report
successful exit, and carry on.  That works, stubbing the instance, to make
this both fairly robust and entirely portable.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>