(#2244) Plugin mount return a directory if modulepath is valid but has no plugins
Prior to this if there were no plugins in the modulepath then the plugin mount
(used for pluginsync) would return nil. If the modulepath is valid then the
correct behavior would be to appear as an empty directory, just as a normal
file source would. We can hack this by returning the modulepath and turning
off recursion so it appears empty.

The error message one would see previous is:
Could not evaluate: Could not retrieve information from environment production source(s) puppet://host/plugins