(#12386) Disable puppet kick on windows
Previously, running puppet kick on Windows would go into an infinite
loop, because the win32-process version of Process.wait returns nil
instead of Errno::ECHILD, if there are no child processes. This was
causing puppet kick to go into an infinite loop.

This commit disables the puppet kick application on Windows, in the
same way that we handle running puppet master on Windows. I also
audited the code base to ensure Process.wait is not being used
elsewhere, and it's not.