(#11535) Platform agnosticism: Desc strings should refer to path separator instead of ':'
We were interpolating the File::PATH_SEPARATOR character in the desc string to
display the actual character being used on the platform on which it is
printed. This is marginally useful when running with --genconfig, but since
the config reference properly lives online, it should be platform-agnostic for
the benefit of Windows users.