(#11535) Call out differing config defaults on Windows
The default values of several configuration settings (diff, color, daemonize)
are different on Windows, but these differences did not appear in the online
configuration reference, as it is generated on a POSIX system. This commit
adds notes on these differing defaults to the description strings of the
affected settings, to ensure that the configuration reference remains
accurate.