(#12399) DRY up rest.rb
Move comment to commit message.

Encapsulate call to request.do_request with the arguments from this class
Then yield to the code block that was called in

We certainly could have retained the full request.do_request(...) {|r| ... }
but this makes the code much cleaner and we only then actually make the call
to request.do_request from here, thus if we change what we pass or how we
get it, we only need to change it here.