(#12359) Improve logging of backtraces
There were a ton of places in the code where we were doing this:

    rescue => detail
       puts detail.backtrace if Puppet[:trace]

In addition to this not being very DRY, it was making it harder to debug the puppet master from within the acceptance testing framework (because when puppet is run in a daemon mode, stdout is swallowed... the "puts" messages were not showing up in the log files or anywhere else, so it was impossible to see the backtrace when something went wrong on the master).

This commit refactors all(?) occurrences of this pattern to call a new utility method:  Puppet::Util::Logging.log_exception(), which provides one single code path for handling exception logging.  It still only logs the backtrace if Puppet[:trace] == true, but it now logs it through the standard Puppet logging framework rather than via "puts".