(#12412) Skip ca acceptance test on Window agents
Previously, this test would fail on Windows agents, because the ca
code makes assumptions about posix style paths. Since this
functionality is not supported on Windows anyways, this commit just
skips the test when run on Windows agents. In order to make that
possible, it refactors the test to use the `agents.each do |agent|`
style test.