(#12349) Order stderr output to fix acceptance test for `module list`
The stdout was ordered, but not the stderr, so the acceptance tests that
passed on my system failed when run by CI.

To order the stderr output I changed the format of the method that gets
unmet dependencies to be a hash that contains the dependency module name
in addition to the error message so that I could sort by that name.

There's also some additional documentation updates to the list command
to show example error output for unmet dependencies.

I changed the unsatisfied_dependencies method to unmet_dependencies
since in future changes that's how dependency problems will be described
to the user, and it's shorter.