(#12126) Refactor Puppet::Util::Autoload
This consolidates all the the loading logic into the singleton associated with
the Autoload class. The Autoload instance is now a thin proxy to this
singleton. This may not seem like an improvement in the short term but it lays
the groundwork for reloading support. Having the logic in the singleton gives
load methods access to the full tree of loaded files, which will be necessary
when we want to reload all changed files.