(#12339) Fix a test failure in 1.8.5.
Hash::[] is documented to take a list of arguments that has an even length, and create
a new Hash instance from that list (as key-value pairs).  Ruby 1.8.7 (and probably 1.8.6)
seem willing to also accept an array of two-element arrays as a compromise; unfortunately
this behavior is not portable.  Splatting the flattened array is an effective workaround.

Similarly, Enumerable#sort_by is provided by 1.8.5, but #max_by is not.