Facter 1.5.7 compatibility in the test condition.
The `augeasversion` fact was only introduced in Facter 1.6.1, and this test
depends on it - or blows up trying to process `nil` in `versioncmp`.

If we don't have the fact we might as well skip over the test set; it isn't
much worse than working harder at this, and less false positives than guessing
the other way around.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>