(#12412) Don't assume the `root` user exists
Previously, the test was verifying that puppet didn't create the
`root` user, assuming it already existed on the host. But that's not
the case on Windows.

This commit just changes the test to use `Host['user']`, which always
refers to the user that we ssh'ed to the host as, and therefore,
always exists.