(#12258) PMT uninstall command should check for metadata
Before this commit the uninstall action does not check if a module has
metadata before uninstalling the module. The uninstall action also
permits using the module directory name (apache) instead of the module
full-name (puppetlabs-apache) when specifying a module to uninstall.

This patch resolves this issue by checking the module has_metadata?
attribute before uninstalling the module.

This patch also enforces the use of only the module full-name when
specifying a module to uninstall:

    $ puppet module uninstall author-modulename

This patch also includes updated tests for the changes in behavior.