(maint) The list action should use the module full name
Before this patch the list action uses the folder name of the module
instead of the full module name when rendering output to the console.
This causes confusion to the end user since modules are installed and
uninstalled using the module full name (puppetlabs-apache) and not the
folder name (apache)

This patch fixes this issue by using the module full name when rendering
to the console.

This patch includes updated specs for the change in behavior.