Clarify interactions between install dir and modulepath options
Having a separate dir options can be a bit confusing since dependencies
are validated for the whole modulepath, so you need to also specify the
modulepath if you want the module being installed to do validation
against other places.  If you don't specify modulepath, dir is assumed
to be the only directory in the modulepath.