(#11858) When syncing plugins, use autoloader to reload changes
Before this, all files synced would be loaded. This resulted in inconsistent
behavior, since not all files would necessarily be loaded during runs when
they did not need to be synced. This makes it so that any already loaded files
that are changed by pluginsync will get reloaded, and any other files are left
to be autoloaded as they normally would be.