(#12564) Paste execpipe commands together with spaces.
Previously the `execpipe` method would create a command from an array by
pasting the content together as strings.  This was all well and good, but it
required the caller to add whitespace manually when invoking the command.

Instead, we can paste together with a space character - leading to a better
external experience, and less repetitious code.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>