Moving the depenency resolution code back into the installer
Having code in the forge.rb that never talked to the forge was awkward,
mainly because it required passing all the install options down to the
forge code.

I went to deal with the force option, and ended up passing yet another
options down to levels that should never have needed to be passed
anything.

Also converting the forge class to a module.  It had an initializer that
was never passed a url, except in tests.